- restructured directory structure
